third_party/blink/web_tests/platform/generic/fast/events/touch/gesture/focus-selectionchange-on-tap-expected.txt

Verifies that a GestureTap triggers focus and selectionchange if and only if mousedown isn't prevented.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Tap on input field but consume mousedown
Sending GestureTapDown
Sending GestureShowPress
Sending GestureTap
Received mouseover on target
Received mousemove on target
Received mousedown on target
Received mouseup on target
Received click on target
PASS tapHandled is true
PASS isFocused(target) is false


Tap on input field without consuming mousedown
Sending GestureTapDown
Sending GestureShowPress
Sending GestureTap
Received mousemove on target
Received mousedown on target
Received mouseup on target
Received click on target
PASS tapHandled is false
Received selectionchange on #document anchor=BODY[2]
PASS isFocused(target) is true


Tap elsewhere to clear focus
Sending GestureTapDown
Sending GestureShowPress
Sending GestureTap
Received mouseout on target
PASS tapHandled is false
Received selectionchange on #document anchor=#text[0]
PASS isFocused(target) is false


PASS successfullyParsed is true

TEST COMPLETE
